services/embedding-engine/requirements.txt
sentence-transformers==2.2.2
torch==2.0.1
scikit-learn==1.3.0
numpy==1.24.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
flask==2.3.3 